refactor(brutus): extract output formatting to output.go
Move 3 output functions (human, valid-only, JSONL).
Part of main.go refactoring (7/9).

Related: C1/C2 (split 1340-line main.go)
